% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BAMBA-package.R
\docType{package}
\name{BAMBA-package}
\alias{BAMBA-package}
\title{BAMBA (Bayesian Analysis of Multiplexed Binding Antibody data)}
\description{
This package fits a Bayesian grouped mixture model for BAMA and
Fc array data using stan. The model identifies responses in the
data.
}
\seealso{
\itemize{
    \item \code{\link{BAMBA}}, for the main model fitting routine.
  }
}
