% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BAMBAResult-methods.r
\name{print.BAMBAResult}
\alias{print.BAMBAResult}
\title{Print a BAMBAResult Object}
\usage{
\method{print}{BAMBAResult}(x, ...)
}
\arguments{
\item{x}{An object of class \code{BAMBAResult}.}

\item{...}{Optional arguments; currently unused.}
}
\description{
This function prints basic information about the model fit by a
\code{\link{BAMBA}} call.
}
\examples{
print(bambaRes)
}
