% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BAMBAResult-methods.r
\name{summary.BAMBAResult}
\alias{summary.BAMBAResult}
\title{Summarize a BAMBAResult Object}
\usage{
summary.BAMBAResult(object, ...)
}
\arguments{
\item{object}{An object of class \code{BAMBAResult}.}

\item{...}{Optional arguments; currently unused.}
}
\description{
This function prints basic information about the model fit by a
\code{\link{BAMBA}} call.
}
\examples{
summary(bambaRes)
}
